Update radius property type to double in location models and adjust related logic in reports map

+++ backend/Models/GetReportsLocation.cs
@@ -4,6 +4,6 @@
     {
         public decimal Latitude { get; set; }
         public decimal Longitude { get; set; }
-        public int Radius { get; set; }
+        public double Radius { get; set; }
     }
 }
